+++ http4k-security-oauth/src/main/kotlin/org/http4k/security/AccessToken.kt
@@ -3,11 +3,11 @@ package org.http4k.security
 import com.fasterxml.jackson.annotation.JsonProperty
 import org.http4k.core.Body
 import org.http4k.format.Jackson.auto
-import org.http4k.security.openid.IdTokenContainer
+import org.http4k.security.openid.IdToken
 
-data class AccessTokenContainer(val value: String)
+data class AccessToken(val value: String)
 
-data class AccessTokenDetails(val accessToken: AccessTokenContainer, val idToken: IdTokenContainer? = null)
+data class AccessTokenDetails(val accessToken: AccessToken, val idToken: IdToken? = null)
 
 data class AccessTokenResponse(
     @JsonProperty("access_token") val accessToken: String,