+++ http4k-security-oauth/src/main/kotlin/org/http4k/security/oauth/server/AccessTokens.kt
@@ -1,7 +1,7 @@
 package org.http4k.security.oauth.server
 
 import com.natpryce.Result
-import org.http4k.security.AccessTokenContainer
+import org.http4k.security.AccessToken
 
 /**
  * Provides a consistent way to generate access tokens.
@@ -10,10 +10,10 @@ interface AccessTokens {
     /**
      * Creates a new access token for a valid authorization code.
      */
-    fun create(authorizationCode: AuthorizationCode): Result<AccessTokenContainer, AuthorizationCodeAlreadyUsed>
+    fun create(authorizationCode: AuthorizationCode): Result<AccessToken, AuthorizationCodeAlreadyUsed>
 
     /**
      * creates a new access token for a given client.
      */
-    fun create(clientId: ClientId): Result<AccessTokenContainer, AccessTokenError>
+    fun create(clientId: ClientId): Result<AccessToken, AccessTokenError>
 }