+++ http4k-security-oauth/src/main/kotlin/org/http4k/security/oauth/server/deprecatedOAuth.kt
@@ -1,0 +1,6 @@
+package org.http4k.security.oauth.server
+
+import org.http4k.security.AccessToken
+
+@Deprecated("use renamed version", ReplaceWith("org.http4k.security.oauth.server.AccessToken"))
+typealias AccessTokenContainer = AccessToken