+++ http4k-security-oauth/src/test/kotlin/org/http4k/security/FakeOAuthPersistence.kt
@@ -6,7 +6,7 @@ import org.http4k.core.Response
 class FakeOAuthPersistence : OAuthPersistence {
 
     var csrf: CrossSiteRequestForgeryToken? = null
-    var accessToken: AccessTokenContainer? = null
+    var accessToken: AccessToken? = null
 
     override fun retrieveCsrf(request: Request): CrossSiteRequestForgeryToken? = csrf
 
@@ -15,9 +15,9 @@ class FakeOAuthPersistence : OAuthPersistence {
         return redirect.header("action", "assignCsrf")
     }
 
-    override fun retrieveToken(request: Request): AccessTokenContainer? = accessToken
+    override fun retrieveToken(request: Request): AccessToken? = accessToken
 
-    override fun assignToken(request: Request, redirect: Response, accessToken: AccessTokenContainer): Response {
+    override fun assignToken(request: Request, redirect: Response, accessToken: AccessToken): Response {
         this.accessToken = accessToken
         return redirect.header("action", "assignToken")
     }